# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "4.3.0",
    "title": "Job's Portal",
    "description": "You get best here",
    "license": {
      "name": "Sukhman",
      "url": "https://epitometechnologies.org/"
    }
  },
  "host": "localhost:8000",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
  "/user/job/org/{id}": {
    "get": {
      "summary": "All Jobs from one ORG",
      "tags": [
        "Users API"
      ],
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of ORG to find All Jobs",
          "type": "string"
        }
      ],
      "responses": {
        "200": {
          "description": "ORG Jobs"
        }
      }
    
  }
},
...
      "/org/signup": {
      "post": {
        "tags": [
          "Organization API"
        ],
        "summary": "Sign UP",
        "description": "Sign new Org in system",
        "parameters": [
          {
            "name": "Organization",
            "in": "body",
            "type": "object",
            "description": "Insert Org data",
            "schema": {
              "$ref": "#/definitions/org"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New Organization is created"
          }
        }
      }
    },
...
    "/org/login": {
      "post": {
        "tags": [
          "Organization API"
        ],
        "summary": "Login",
        "description": "login  in system",
        "parameters": [
          {
            "name": "Organization",
            "in": "body",
            "type": "object",
            "description": "Login Organization ",
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Logged In"
          }
        }
      }
    },
    "/org/forget_pass": {
      "post": {
        "tags": [
          "Organization API"
        ],
        "summary": "Forget Password",
        "description": "forget pass   of ORG",
        "parameters": [
          {
            "name": "Organization",
            "in": "body",
            "type": "object",
            "description": "Forget Pass ",
            "schema": {
              "$ref": "#/definitions/Forget_password"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "token sent  Succesfully"
          }
        }
      }
    },
    "/org/reset_pass/{id}/{token}": {
      "post": {
        "tags": [
          "Organization API"
        ],
        "summary": "Reset Password",
        "description": "Reset pass   of ORG",
        "parameters": [
          {
            "name": "id",
            "in": "path", 
            "description": "Insert Id "
          },
          {
            "name": "token",
            "in": "path", 
            "description": "Insert token" 
          },
          {
            "name": "Organization",
            "in": "body",
            "type": "object",
            "description": "Insert new Password ",
            "schema": {
              "$ref": "#/definitions/Reset_password"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": { 
            "description": "Password update Succesfully"
          }
        }
      }
    },
    "/org/{id}": {
      "get": {
        "summary": "ORG with given ID",
        "tags": [
          "Organization API"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of ORG that we want to find",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "ORG detail"
          }
        }
      
    },
    "patch": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of ORG that we want to update",
          "type": "string"
        },
        {
          "name": "ORG",
          "in": "body",
          "type": "object",
          "schema":{
            "$ref":"#/definitions/Update_org"
          }
        }
      ],
      "summary": "ORG data with given ID",
      "tags": [
        "Organization API"
      ],
      "responses": {
        "200": {
          "description": "ORG data"
        }
      }
    },
    "delete": {
      "tags": [
        "Organization API"
      ],
      "summary": "delete ORG data with given ID",
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of ORG that we want to delete",
          "type": "string"
        }
      ],
      "responses": {
        "200": {
          "description": "ORG data"
        }
      }
    }
   },
...
    "/org/updatejob/{id}": {
      "patch": {
        "tags": [
          "Organization API"
        ],
        "summary": "Update Job",
        "description": "Update Job in ORG",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Insert JOb Id"
          },
          { 
            "name": "Job",
            "in": "body", 
            "type": "object",
            "description": "Update Job data",
            "schema":{
              "$ref":"#/definitions/Create_Job"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Job is Updated" 
          }
        }
      }
    },
    "/org/deljob/{id}": {
      "delete": {
        "tags": [
          "Organization API"
        ],
        "summary": "Delete Job",
        "description": "Delete Job in ORG",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": " Insert Job Id"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Job is Deleted" 
          }
        }
      }
    },
...
    "/org/Job/With/Applicant/{id}": {
      "get": {
        "summary": "Job with Applicant's",
        "tags": [
          "Organization API"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of Job that we want to find",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Users Job"
          }
        }
      
    }
  }
  },
  "definitions": {
...
    "Login": {
      "properties":{
        "Email":{
          "type": "string"
        },
        "Password":{
          "type": "string"
        }
      }
    },
    "Forget_password": {
      "properties":{
        "Email":{
          "type": "string"
        }
      }
    },
    "Reset_password": {
      "properties":{
        "Email":{
          "type": "string"
        },
        "Password":{
          "type": "string"
        },
        "ConfirmPassword":{
          "type": "string"
        }
      }
    },
...
     "org": {
       "properties":{
         "Name":{
           "type": "string"
         },
         "Email":{
           "type": "string"
         },
         "Password":{
           "type": "string"
         },
         "Address":{
           "type": "string"
         },
         "Phone":{
           "type": "number"
         },
         "Country":{
           "type": "string"
         } 
       }
     } ,
     "Update_org":{
       "properties":{
        "Name":{
          "type": "string"
        },
        "Address":{
          "type": "string"
        },
        "Phone":{
          "type": "number"
        },
        "Country":{
          "type": "string"
        } 
       }
     },
     "Create_Job":{
      "properties":{
        "JobTitle":{
          "type": "string"
        },
        "Description":{
          "type": "string"
        },
        "RequiredSkills":{
          "type": "array",
          "items":{
            "type": "string"
          }
        },
        "Salary":{
          "type": "string"
        },
        "Location":{
          "type": "string"
        } 
       }
     }
  }
  }